Always replace the default 'wodby.yml'
We usually don't do any changes to the post-deployment script 'wodby.yml', therefore
i suggest that we actually ignore this file as we do for many others, and always use the
most updated one. That way if we update it in this repo also all sites will get the most
recent version.
This can still easily be overridden on a per-project basis.