Add "permission fix" to post-deployment wodby.yml
By adding this step we try to prevent errors on the post-deployment action which will leave the site on a unreliable state.

The `drush deploy` action will try to remove the aggregated css/js as part of the cache-rebuild step, because those files are often generated by the php user `www-data` and the post-deploy is handled by the ssh user `wodby` this can lead to permission mismatches and stops the whole post-depoly script. By adding this initial step we mitigate this error while adding minimal overhead.